Starting to work out how to structure the code using multiple crates
