+++ spritec-preview/README.md
@@ -1,0 +1,5 @@
+# spritec-preview
+
+This is the GUI frontend to spritec. It's purpose is to provide a convenient way
+to preview the spritesheets you generate with spritec. This tool is great for
+quickly iterating as you update your character's design.
