+++ src/material.rs
@@ -8,7 +8,7 @@ pub struct Material {
 impl Default for Material {
     fn default() -> Self {
         Self {
-            diffuse_color: Rgba::black(),
+            diffuse_color: Rgba {r: 0.0, g: 0.0, b: 0.0, a: 0.0},
         }
     }
 }
