+++ src/shaders.rs
@@ -1,0 +1,5 @@
+mod cel;
+mod outline;
+
+pub use cel::*;
+pub use outline::*;
